+++ packages/common/src/utils/stream-utils.ts
@@ -255,7 +255,7 @@ export class StreamUtils {
     ipfs: IpfsApi
   ): Promise<CeramicCommit> {
     if (StreamUtils.isSignedCommit(commit)) {
-      const block = await ipfs.block.get((commit as DagJWS).link)
+      const block = await ipfs.block.get(toCID((commit as DagJWS).link))
       return {
         jws: commit as DagJWS,
         linkedBlock: block,