+++ packages/core/src/__tests__/ceramic-api.test.ts
@@ -9,6 +9,7 @@ import { createIPFS } from '@ceramicnetwork/ipfs-daemon'
 import { createCeramic } from './create-ceramic.js'
 import { ModelInstanceDocument } from '@ceramicnetwork/stream-model-instance'
 import { Model, ModelDefinition } from '@ceramicnetwork/stream-model'
+import type { AddOperation } from 'fast-json-patch'
 
 /**
  * Generates string of particular size in bytes
@@ -58,6 +59,23 @@ describe('Ceramic API', () => {
     },
   }
 
+  const MODEL_DEFINITION_BLOB: ModelDefinition = {
+    name: 'MyBlobModel',
+    version: Model.VERSION,
+    accountRelation: { type: 'list' },
+    schema: {
+      $schema: 'https://json-schema.org/draft/2020-12/schema',
+      type: 'object',
+      additionalProperties: false,
+      properties: {
+        myData: {
+          type: 'string',
+        },
+      },
+      required: ['myData'],
+    },
+  }
+
   beforeAll(async () => {
     ipfs = await createIPFS()
   })
@@ -71,6 +89,7 @@ describe('Ceramic API', () => {
     let tmpFolder: tmp.DirectoryResult
 
     beforeEach(async () => {
+      ModelInstanceDocument.MAX_DOCUMENT_SIZE = 16_000_000
       tmpFolder = await tmp.dir({ unsafeCleanup: true })
       ceramic = await createCeramic(ipfs, {
         stateStoreDirectory: tmpFolder.path,
@@ -171,7 +190,6 @@ describe('Ceramic API', () => {
     it('can create and update stream with valid model to trigger indexing', async () => {
       const CONTENT0 = { myData: 0 }
       const CONTENT1 = { myData: 1 }
-
       // TODO (NET-1614): Extend with targeted payload comparison
       const addIndexSpy = jest.spyOn(ceramic.repository, 'indexStreamIfNeeded')
       const model = await Model.create(ceramic, MODEL_DEFINITION)
@@ -183,6 +201,68 @@ describe('Ceramic API', () => {
       await doc.replace(CONTENT1)
       expect(doc.content).toEqual(CONTENT1)
       expect(addIndexSpy).toBeCalledTimes(3)
+      addIndexSpy.mockRestore()
+    })
+
+    it('can create and update stream with valid model under size limits', async () => {
+      const CONTENT0 = { myData: 'abcdef' }
+      const CONTENT1 = { myData: 'ghi' }
+      ModelInstanceDocument.MAX_DOCUMENT_SIZE = 30
+      const addIndexSpy = jest.spyOn(ceramic.repository, 'indexStreamIfNeeded')
+      const model = await Model.create(ceramic, MODEL_DEFINITION_BLOB)
+      expect(addIndexSpy).toBeCalledTimes(1)
+      const midMetadata = { model: model.id }
+      const doc = await ModelInstanceDocument.create(ceramic, CONTENT0, midMetadata)
+      expect(doc.content).toEqual(CONTENT0)
+      expect(addIndexSpy).toBeCalledTimes(2)
+      await doc.replace(CONTENT1)
+      expect(doc.content).toEqual(CONTENT1)
+      expect(addIndexSpy).toBeCalledTimes(3)
+      addIndexSpy.mockRestore()
+    })
+
+    it('will fail to create stream over size limits', async () => {
+      const CONTENT0 = { myData: 'abcdefghijklmn' }
+      ModelInstanceDocument.MAX_DOCUMENT_SIZE = 10
+      const addIndexSpy = jest.spyOn(ceramic.repository, 'indexStreamIfNeeded')
+      const model = await Model.create(ceramic, MODEL_DEFINITION_BLOB)
+      expect(addIndexSpy).toBeCalledTimes(1)
+      const midMetadata = { model: model.id }
+      await expect(ModelInstanceDocument.create(ceramic, CONTENT0, midMetadata)).rejects.toThrow(
+        /which exceeds maximum size/
+      )
+      addIndexSpy.mockRestore()
+    })
+
+    it('will update stream if under size limits', async () => {
+      const CONTENT0 = { myData: 'abcdef' }
+      const CONTENT1 = [{ op: 'replace', path: '/myData', value: 'abcdefgh' } as AddOperation]
+      ModelInstanceDocument.MAX_DOCUMENT_SIZE = 30
+      const addIndexSpy = jest.spyOn(ceramic.repository, 'indexStreamIfNeeded')
+      const model = await Model.create(ceramic, MODEL_DEFINITION_BLOB)
+      expect(addIndexSpy).toBeCalledTimes(1)
+      const midMetadata = { model: model.id }
+      const doc = await ModelInstanceDocument.create(ceramic, CONTENT0, midMetadata)
+      expect(doc.content).toEqual(CONTENT0)
+      expect(addIndexSpy).toBeCalledTimes(2)
+      await doc.patch(CONTENT1)
+      expect(doc.content).toEqual({ myData: 'abcdefgh' })
+      expect(addIndexSpy).toBeCalledTimes(3)
+      addIndexSpy.mockRestore()
+    })
+
+    it('will fail to update stream over size limits', async () => {
+      const CONTENT0 = { myData: 'abcdef' }
+      const CONTENT1 = [{ op: 'replace', path: '/myData', value: 'abcdefghijkl' } as AddOperation]
+      ModelInstanceDocument.MAX_DOCUMENT_SIZE = 20
+      const addIndexSpy = jest.spyOn(ceramic.repository, 'indexStreamIfNeeded')
+      const model = await Model.create(ceramic, MODEL_DEFINITION_BLOB)
+      expect(addIndexSpy).toBeCalledTimes(1)
+      const midMetadata = { model: model.id }
+      const doc = await ModelInstanceDocument.create(ceramic, CONTENT0, midMetadata)
+      expect(doc.content).toEqual(CONTENT0)
+      expect(addIndexSpy).toBeCalledTimes(2)
+      await expect(doc.patch(CONTENT1)).rejects.toThrow(/which exceeds maximum size/)
       addIndexSpy.mockRestore()
     })