+++ packages/core/src/__tests__/create-ceramic.ts
@@ -28,6 +28,7 @@ export async function createCeramic(
       db: databaseUrl.href,
       allowQueriesBeforeHistoricalSync: false,
       disableComposedb: false,
+      enableHistoricalSync: false,
     },
     ...config,
   }