+++ packages/core/src/ceramic.ts
@@ -106,7 +106,6 @@ const DEFAULT_APPLY_COMMIT_OPTS = {
 const DEFAULT_CREATE_FROM_GENESIS_OPTS = {
   anchor: true,
   publish: true,
-  pin: true,
   sync: SyncOptions.PREFER_CACHE,
   ...DEFAULT_CLIENT_INITIATED_WRITE_OPTS,
 }
@@ -131,8 +130,6 @@ export interface CeramicConfig {
   gateway?: boolean
 
   indexing?: IndexingConfig
-  // TODO: Replace in CDB-2072
-  sync?: boolean
 
   networkName?: string
   pubsubTopic?: string
@@ -539,18 +536,13 @@ export class Ceramic implements CeramicApi {
       maxQueriesPerSecond
     )
 
-    const sync =
-      config.sync == undefined
-        ? process.env.CERAMIC_ENABLE_EXPERIMENTAL_SYNC === 'true'
-        : config.sync
-
     const params: CeramicParameters = {
       gateway: config.gateway,
       stateStoreDirectory: config.stateStoreDirectory,
       indexingConfig: config.indexing,
       networkOptions,
       loadOptsOverride,
-      sync,
+      sync: config.indexing?.enableHistoricalSync,
     }
 
     const modules: CeramicModules = {