+++ packages/core/src/indexing/index-query-not-available.error.ts
@@ -2,6 +2,8 @@ import type { StreamID } from '@ceramicnetwork/streamid'
 
 export class IndexQueryNotAvailableError extends Error {
   constructor(model: StreamID | string) {
-    super(`Index for historical data on ${model} is not available`)
+    super(
+      `Queries on Model ${model} are not currently available because historical data for that model is still syncing`
+    )
   }
 }