+++ packages/cli/CHANGELOG.md
@@ -3,6 +3,87 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.28.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/cli@2.28.0-rc.0...@ceramicnetwork/cli@2.28.0) (2023-02-28)
+
+**Note:** Version bump only for package @ceramicnetwork/cli
+
+
+
+
+
+# [2.28.0-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/cli@2.27.0...@ceramicnetwork/cli@2.28.0-rc.0) (2023-02-28)
+
+
+### Bug Fixes
+
+* **cli:** legacy pin, must return json for 200 ([#2769](https://github.com/ceramicnetwork/js-ceramic/issues/2769) ([e6124eb](https://github.com/ceramicnetwork/js-ceramic/commit/e6124ebe179d75a19993947f4a4f6626214ffced))
+
+
+
+
+
+# [2.27.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/cli@2.27.0-rc.0...@ceramicnetwork/cli@2.27.0) (2023-02-28)
+
+**Note:** Version bump only for package @ceramicnetwork/cli
+
+
+
+
+
+# [2.27.0-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/cli@2.26.0...@ceramicnetwork/cli@2.27.0-rc.0) (2023-02-28)
+
+**Note:** Version bump only for package @ceramicnetwork/cli
+
+
+
+
+
+# [2.26.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/cli@2.26.0-rc.0...@ceramicnetwork/cli@2.26.0) (2023-02-28)
+
+**Note:** Version bump only for package @ceramicnetwork/cli
+
+
+
+
+
+# [2.26.0-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/cli@2.25.0-rc.0...@ceramicnetwork/cli@2.26.0-rc.0) (2023-02-28)
+
+
+### Bug Fixes
+
+* **cli:** legacy pin api, empty 200 on add to support 3id-connect, ot… ([#2762](https://github.com/ceramicnetwork/js-ceramic/issues/2762)) ([d5f076a](https://github.com/ceramicnetwork/js-ceramic/commit/d5f076a37f3a4b0367c796277e56e3ce0b9cb355))
+
+
+
+
+
+# [2.25.0-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/cli@2.24.0-rc.0...@ceramicnetwork/cli@2.25.0-rc.0) (2023-02-27)
+
+
+### Features
+
+* **cli, core:** add sync worker flag to config file with default values ([#2757](https://github.com/ceramicnetwork/js-ceramic/issues/2757)) ([5375c88](https://github.com/ceramicnetwork/js-ceramic/commit/5375c886ecf58fa646ab3afe0796b0d2edc17af4))
+
+
+
+
+
+# [2.24.0-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/cli@2.23.0-rc.0...@ceramicnetwork/cli@2.24.0-rc.0) (2023-02-24)
+
+
+### Bug Fixes
+
+* **cli:** validate daemon config after applying env overrides ([#2747](https://github.com/ceramicnetwork/js-ceramic/issues/2747)) ([89ffe58](https://github.com/ceramicnetwork/js-ceramic/commit/89ffe583350bb4bc3d9f9977ddd71dbb8804f58d))
+
+
+### Features
+
+* Bump IPFS/IPLD deps ([#2746](https://github.com/ceramicnetwork/js-ceramic/issues/2746)) ([ef23e50](https://github.com/ceramicnetwork/js-ceramic/commit/ef23e509556f32e6b1f6c1ed6f87116a3bc7e26a))
+
+
+
+
+
 # [2.23.0-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/cli@2.22.0...@ceramicnetwork/cli@2.23.0-rc.0) (2023-02-23)