+++ packages/ipfs-daemon/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.16.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/ipfs-daemon@2.16.0-rc.0...@ceramicnetwork/ipfs-daemon@2.16.0) (2023-02-28)
+
+**Note:** Version bump only for package @ceramicnetwork/ipfs-daemon
+
+
+
+
+
+# [2.16.0-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/ipfs-daemon@2.15.0-rc.0...@ceramicnetwork/ipfs-daemon@2.16.0-rc.0) (2023-02-27)
+
+
+### Features
+
+* add pubsub cache ttl and bootstrap peers to bundled kubo config ([#2756](https://github.com/ceramicnetwork/js-ceramic/issues/2756)) ([d720074](https://github.com/ceramicnetwork/js-ceramic/commit/d7200747c069b8211a1c7f4ca18903b2838bdff1))
+
+
+
+
+
+# [2.15.0-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/ipfs-daemon@2.14.0-rc.0...@ceramicnetwork/ipfs-daemon@2.15.0-rc.0) (2023-02-24)
+
+
+### Features
+
+* Bump IPFS/IPLD deps ([#2746](https://github.com/ceramicnetwork/js-ceramic/issues/2746)) ([ef23e50](https://github.com/ceramicnetwork/js-ceramic/commit/ef23e509556f32e6b1f6c1ed6f87116a3bc7e26a))
+
+
+
+
+
 # [2.14.0-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/ipfs-daemon@2.13.0...@ceramicnetwork/ipfs-daemon@2.14.0-rc.0) (2023-02-23)